Changed capitalization of the package Model
