+++ src/main/java/ar/fiuba/tdd/template/model/CellContent.java
@@ -1,4 +1,4 @@
-package ar.fiuba.tdd.template.Model;
+package ar.fiuba.tdd.template.model;
 
 public abstract class CellContent<T> {
