+++ src/main/java/ar/fiuba/tdd/template/model/ValueContent.java
@@ -1,11 +1,8 @@
-package ar.fiuba.tdd.template.Model;
+package ar.fiuba.tdd.template.model;
 
+public class ValueContent<T> extends CellContent<T> {
 
-public class ClueContent<T> extends CellContent<T> {
-
-    private T value;
-
-    public ClueContent(T value) {
+    public ValueContent(T value) {
         this.value = value;
     }
